requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Babel==2.14.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0
django-environ==0.11.2
django-money==3.4.1
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
drf-yasg==1.21.7
GDAL==3.10.2
gis==0.2.1
hiredis==2.3.2
idna==3.6
inflection==0.5.1
packaging==24.0
phonenumberslite==8.13.31
pillow==10.2.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
py-moneyed==3.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.1
